add autenticazione per admin e aggiunta voce admin

+++ src/routers/products.ts
@@ -1,12 +1,16 @@
 import { createClient } from "@vercel/postgres";
 import express, { Response, Request } from "express";
 import { config } from "dotenv";
-import "dotenv/config";
-import { authenticateToken } from "../JWT/authenticateToken";
+import { authenticateToken, JwtRequest } from "../JWT/authenticateToken";
 export const routerProducts = express.Router();
+
+config();
+
 const client = createClient({
   connectionString: process.env.DATABASE_URL,
 });
+
+client.connect();
 
 routerProducts.get("", async (req: Request, res: Response) => {
   const products = await client.query("SELECT * FROM products");
@@ -14,12 +18,10 @@ routerProducts.get("", async (req: Request, res: Response) => {
 });
 
 // create
-routerProducts.post(
-  "",
-  authenticateToken,
-  (req: Request, res: Response) => {
-    const { title, price, category, description, image } = req.body;
-
+routerProducts.post("", authenticateToken, (req: JwtRequest, res: Response) => {
+  const { title, price, category, description, image } = req.body;
+  const access = req.user as { id: number; admin: boolean };
+  if (access.admin === true) {
     client.query(
       "INSERT INTO products ( title, price, category, description, image) VALUES ($1, $2, $3, $4, $5) RETURNING *",
       [title, price, category, description, image],
@@ -28,8 +30,9 @@ routerProducts.post(
         else return res.status(200).json(result.rows);
       }
     );
-  }
-);
+    res.status(200).json({ message: "Avvenuta con successo" });
+  } else res.status(400).json({ message: "Non sei autorizzato" });
+});
 
 // reade
 routerProducts.get("/:id", (req: Request, res: Response) => {
@@ -49,19 +52,21 @@ routerProducts.get("/:id", (req: Request, res: Response) => {
 routerProducts.put(
   "/:id",
   authenticateToken,
-  (req: Request, res: Response) => {
+  (req: JwtRequest, res: Response) => {
     const productId = req.params.id;
-
     const { title, price, category, description, image } = req.body;
-
-    client.query(
-      "UPDATE products SET title = $1, price = $2, category = $3, description = $4, image = $5 WHERE id = $6 RETURNING *",
-      [title, price, category, description, image, productId],
-      (err, result) => {
-        if (err) return res.status(400).json({ err });
-        else return res.status(200).json(result.rows);
-      }
-    );
+    const access = req.user as { id: number; admin: boolean };
+    if (access.admin === true) {
+      client.query(
+        "UPDATE products SET title = $1, price = $2, category = $3, description = $4, image = $5 WHERE id = $6 RETURNING *",
+        [title, price, category, description, image, productId],
+        (err, result) => {
+          if (err) return res.status(400).json({ err });
+          else return res.status(200).json(result.rows);
+        }
+      );
+      res.status(200).json({ message: "Avvenuta con successo" });
+    } else res.status(400).json({ message: "Non sei autorizzato" });
   }
 );
 
@@ -69,16 +74,19 @@ routerProducts.put(
 routerProducts.delete(
   "/:id",
   authenticateToken,
-  (req: Request, res: Response) => {
+  (req: JwtRequest, res: Response) => {
     const productId = req.params.id;
-
-    client.query(
-      "DELETE FROM products WHERE id = $1",
-      [productId],
-      (err, result) => {
-        if (err) return res.status(400).json({ err });
-        else return res.status(200).json(result.rows);
-      }
-    );
+    const access = req.user as { id: number; admin: boolean };
+    if (access.admin === true) {
+      client.query(
+        "DELETE FROM products WHERE id = $1",
+        [productId],
+        (err, result) => {
+          if (err) return res.status(400).json({ err });
+          else return res.status(200).json(result.rows);
+        }
+      );
+      res.status(200).json({ message: "Avvenuta con successo" });
+    } else res.status(400).json({ message: "Non sei autorizzato" });
   }
 );
